requirements.txt
# Main packages
configparserc~=2.0.2
Markdown~=3.6
django-environ~=0.11.2

# REST API packages
djangorestframework~=3.15.2
drf-yasg==1.21.7
django-filter==24.3
drf_orjson_renderer==1.7.1
ormsgpack~=1.5.0
pyyaml~=6.0.2

# web server
uvicorn~=0.30.5
pyuwsgi==2.0.23.post0
# Restore it if some problems with pyuwsgi
# uwsgi==2.0.23
fastapi-slim~=0.112.0
aiofiles~=24.1.0
asgiref>=3.8.1

# Notifications
cent~=5.0.0

# Auth
Authlib~=1.3.1

# Minificators
jsmin~=3.0.1
django-htmlmin~=0.11.0

# 2fa
pyotp~=2.9.0

# Storage
django-storages[libcloud]==1.14.4
